requirements.txt
appnope==0.1.0
astroid==2.3.3
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
gw-phenom==1.0.1
h5py==2.10.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
joblib==0.14.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.1
mccabe==0.6.1
mpmath==1.1.0
numpy==1.17.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkginfo==1.5.0.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
scikit-learn==0.21.3
scipy==1.3.2
six==1.12.0
sympy==1.4
tqdm==4.36.1
traitlets==4.3.3
twine==2.0.0
typed-ast==1.4.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
